// apps/ui-kit/src/lib/components/atoms/button/button.enums.ts
// Copyright (c) 2024 IOTA Stiftung
// SPDX-License-Identifier: Apache-2.0

export enum ButtonSize {
    Small = 'small',
    Medium = 'medium',
}

export enum ButtonType {
    Primary = 'primary',
    Secondary = 'secondary',
    Ghost = 'ghost',
    Outlined = 'outlined',
    Destructive = 'destructive',
}